Move inter-Node types to its own place
