+++ packages/node/src/node.ts
@@ -4,16 +4,13 @@ import {
   Opcode,
   ProtocolMessage
 } from "@counterfactual/machine";
-import {
-  Address,
-  NetworkContext,
-  Node as NodeTypes
-} from "@counterfactual/types";
+import { NetworkContext, Node as NodeTypes } from "@counterfactual/types";
 import { SigningKey } from "ethers/utils";
 import EventEmitter from "eventemitter3";
 
 import { RequestHandler } from "./request-handler";
 import { IMessagingService, IStoreService } from "./services";
+import { NodeMessage } from "./types";
 
 export interface NodeConfig {
   // The prefix for any keys used in the store by this Node depends on the
@@ -153,13 +150,4 @@ export class Node {
     }
     await this.requestHandler.callEvent(msg.event, msg);
   }
-}
-
-/**
- * The message interface for Nodes to communicate with each other.
- */
-export interface NodeMessage {
-  from?: Address;
-  event: NodeTypes.EventName;
-  data: any;
 }
